+++ src/listeners/panel/TaxonomyListener.ts
@@ -5,17 +5,13 @@ import { authentication, window } from 'vscode';
 import { ArticleHelper, Extension, Settings, TaxonomyHelper } from '../../helpers';
 import { BlockFieldData, CustomTaxonomyData, PostMessageData, TaxonomyType } from '../../models';
 import { DataListener } from '.';
-import {
-  DefaultFields,
-  SETTING_SEO_DESCRIPTION_FIELD,
-  SETTING_SEO_TITLE_FIELD
-} from '../../constants';
 import { SponsorAi } from '../../services/SponsorAI';
 import { PanelProvider } from '../../panelWebView/PanelProvider';
 import { MessageHandlerData } from '@estruyf/vscode';
 import * as l10n from '@vscode/l10n';
 import { LocalizationKey } from '../../localization';
 import { Copilot } from '../../services/Copilot';
+import { getDescriptionField, getTitleField } from '../../utils';
 
 export class TaxonomyListener extends BaseListener {
   /**
@@ -73,10 +69,9 @@ export class TaxonomyListener extends BaseListener {
     }
   }
 
-
   /**
    * Suggests a taxonomy for a given command, request ID, and tag type.
-   * 
+   *
    * @param command - The command to execute.
    * @param requestId - The ID of the request.
    * @param type - The type of the tag.
@@ -100,8 +95,8 @@ export class TaxonomyListener extends BaseListener {
     const extPath = Extension.getInstance().extensionPath;
     const panel = PanelProvider.getInstance(extPath);
 
-    const titleField = (Settings.get(SETTING_SEO_TITLE_FIELD) as string) || DefaultFields.Title;
-    const descriptionField = (Settings.get(SETTING_SEO_DESCRIPTION_FIELD) as string) || DefaultFields.Description;
+    const titleField = getTitleField();
+    const descriptionField = getDescriptionField();
 
     const tags = await Copilot.suggestTaxonomy(
       articleDetails.data[titleField],
@@ -127,7 +122,7 @@ export class TaxonomyListener extends BaseListener {
 
   /**
    * Suggests taxonomy based on the provided command, request ID, and tag type.
-   * 
+   *
    * @param command - The command to execute.
    * @param requestId - The ID of the request.
    * @param type - The type of tag.
@@ -166,9 +161,8 @@ export class TaxonomyListener extends BaseListener {
       return;
     }
 
-    const titleField = (Settings.get(SETTING_SEO_TITLE_FIELD) as string) || DefaultFields.Title;
-    const descriptionField =
-      (Settings.get(SETTING_SEO_DESCRIPTION_FIELD) as string) || DefaultFields.Description;
+    const titleField = getTitleField();
+    const descriptionField = getDescriptionField();
 
     const suggestions = await SponsorAi.getTaxonomySuggestions(
       githubAuth.accessToken,